# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Qubah Learning API",
        "description": "REST API for the Qubah Learning educational platform. Provides endpoints for authentication, subjects, topics, content, progress tracking, and parent dashboard.",
        "contact": {
            "name": "Qubah Team",
            "email": "support@qubah.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/lesson-files": {
            "get": {
                "tags": [
                    "Lesson Files"
                ],
                "summary": "Get all lesson files",
                "operationId": "7f45ae30bf0484829e0f849bb8a3f870",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lesson Files"
                ],
                "summary": "Create a new lesson file",
                "operationId": "05621cb90d3ffd0136528a8d48676b7b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "lesson_id",
                                    "type",
                                    "file_path"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "ملف الفيديو الرئيسي"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "video"
                                    },
                                    "file_path": {
                                        "type": "string",
                                        "example": "files/lesson1.mp4"
                                    },
                                    "lesson_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "thumbnail_path": {
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "metadata": {
                                        "type": "object",
                                        "nullable": true
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/lesson-files/{id}": {
            "get": {
                "tags": [
                    "Lesson Files"
                ],
                "summary": "Get a specific lesson file",
                "operationId": "c629e804dda74f14c1e278097354b3a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Lesson Files"
                ],
                "summary": "Update a lesson file",
                "operationId": "35b27e3bbc0713524ae57261120600e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "file_path": {
                                        "type": "string"
                                    },
                                    "lesson_id": {
                                        "type": "integer"
                                    },
                                    "thumbnail_path": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "metadata": {
                                        "type": "object",
                                        "nullable": true
                                    },
                                    "order": {
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Lesson file not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lesson Files"
                ],
                "summary": "Delete a lesson file",
                "operationId": "e85bd589469271d92867e1fdb8ecc16b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Lesson file not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/lesson-files/upload": {
            "post": {
                "tags": [
                    "Lesson Files"
                ],
                "summary": "Upload a new lesson file",
                "operationId": "3ce3592ab43282b1bcd532e0027a1de0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "lesson_id",
                                    "title",
                                    "type"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "lesson_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "400": {
                        "description": "No file provided"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Login via /api/v1/auth/login and use the returned access_token here.",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}